auth ok

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TasksStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,12 +14,12 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTask(filterDTO: GetTasksFilterDTO): Promise<Task[]> {
-    return this.taskRepository.getTask(filterDTO)
+  async getTask(filterDTO: GetTasksFilterDTO, user: User): Promise<Task[]> {
+    return this.taskRepository.getTask(filterDTO, user);
   }
 
-  async getTaskByID(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskByID(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" no found`);
@@ -27,20 +28,20 @@ export class TasksService {
     return found;
   }
 
-  async createTask(createTaskDTO: CreateTaskDTO): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDTO);
+  async createTask(createTaskDTO: CreateTaskDTO, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDTO, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" no found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TasksStatus): Promise<Task> {
-    const task = await this.getTaskByID(id);
+  async updateTaskStatus(id: number, status: TasksStatus, user: User): Promise<Task> {
+    const task = await this.getTaskByID(id, user);
 
     task.status = status;
     await task.save();
